communication/requirements.txt
psycopg2-binary
python-dotenv
pyserial